% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datageneration.R
\name{datageneration}
\alias{datageneration}
\title{Generate artificial balanced panel data by simulation}
\usage{
datageneration(N, S)
}
\arguments{
\item{N}{The number of cross-sectional units}

\item{S}{The length of time series}
}
\value{
A data.frame that contains the following elements.
\item{id}{The unit index}
\item{period}{The time period index}
\item{Y}{The outcome}
\item{D}{The treatment}
\item{X}{The unit-specific covariate}
\item{once}{Whether some treatment realization has been received at least once so far}
\item{event}{The earliest time period of receiving some treatment realization so far (the event date)}
\item{number}{The number of treatment adoptions so far}
}
\description{
The `datageneration()` function generates artificial balanced panel data from
the data generating process considered in the Monte Carlo experiment
in Yanagi (2023).
}
\examples{
set.seed(1)
data <- datageneration(N = 1000, S = 4)

}
\references{
Yanagi, T., 2023.
An effective treatment approach to
difference-in-differences with general treatment patterns.
arXiv:2212.13226.
}
